DB/Creature: Matt at Crystal Lake
Matt is a human boy that lives in the town of Goldshire. He loves to go fishing but appears to not be very good at this due to inexperience. He will walk from the pier at Crystal Lake to the Goldshire pond (which he refers to as his "ol' fishin' hole") attempting to find fish to catch.

(cherry picked from commit 87092140740b7d6e30687345f8df1984222c2946)